Api branch remastered

+++ api/component/models.py
@@ -2,10 +2,12 @@ from bson import ObjectId
 from flask import jsonify, request, abort
 from app import db
 
+
 def convert_object_ids(components):
     for component in components:
         component['_id'] = str(component['_id'])
     return components
+
 
 class Component:
 
@@ -31,7 +33,7 @@ class Component:
                 return jsonify({'status': 'error', 'message': 'Component not found'}), 400
         except Exception as e:
             return jsonify({'status': 'error', 'message': str(e)}), 400
-        
+
     @staticmethod
     def readByPageId():
         try:
@@ -54,12 +56,20 @@ class Component:
     def write():
         try:
             data = request.get_json()
-            current_order = data['order_number']
             if ('order_number' not in data or 'component_type' not in data or 'propTextShort' not in data
                     or 'propTextMid' not in data
                     or 'propTextLong' not in data or 'propImages'
                     not in data or 'visible' not in data):
                 return jsonify({'status': 'error', 'message': 'Missing required fields'}), 400
+
+            existing_component = db['Component'].find_one(
+                {'page_id': data['page_id'], 'order_number': data['order_number']})
+            if existing_component:
+                # If a component already exists at the specified order_number, shift the others
+                db['Component'].update_many({
+                    'order_number': {'$gte': data['order_number']},
+                    'page_id': data['page_id']
+                }, {'$inc': {'order_number': 1}})
 
             result = db['Component'].insert_one({
                 'page_id': data['page_id'],
@@ -72,11 +82,6 @@ class Component:
                 'visible': data['visible']})
 
             if result:
-                db['Component'].update_many({
-                    'order_number': {'$gte': current_order},
-                    'page_id': data['page_id'],
-                    '_id': {'$ne': ObjectId(result.inserted_id)}
-                }, {'$inc': {'order_number': 1}})
                 return jsonify({'status': 'success', 'message': f'Component successfully inserted.'}), 200
             else:
                 return jsonify({'status': 'error', 'message': 'Failed to add component'}), 400
@@ -88,7 +93,15 @@ class Component:
         try:
             update_data = request.get_json()
 
-            current_order = update_data['order_number']
+            existing_component = db['Component'].find_one({'page_id': update_data['page_id'],
+                                                           'order_number': update_data['order_number']})
+            if existing_component:
+                # If a component already exists at the specified order_number, shift the others
+                db['Component'].update_many({
+                    'order_number': {'$gte': update_data['order_number']},
+                    'page_id': update_data['page_id'],
+                    '_id': {'$ne': ObjectId(update_data['_id'])}
+                }, {'$inc': {'order_number': 1}})
 
             result = db['Component'].update_one({'_id': ObjectId(update_data['_id'])}, {'$set': {
                 'page_id': update_data['page_id'],
@@ -101,12 +114,6 @@ class Component:
             }})
 
             if result.modified_count > 0:
-                db['Component'].update_many({
-                    'order_number': {'$gte': current_order},
-                    'page_id': update_data['page_id'],
-                    '_id': {'$ne': ObjectId(update_data['_id'])}
-                }, {'$inc': {'order_number': 1}})
-
                 return jsonify({'status': 'success', 'message': f'Component successfully updated.'}), 200
             else:
                 return jsonify({'status': 'error', 'message': f'Component not found'}), 400
@@ -117,14 +124,21 @@ class Component:
     def delete():
         try:
             data = request.get_json()
-            current_order = request.form.get('order_number')
+
+            component_to_delete = db['Component'].find_one({'_id': ObjectId(data['_id'])})
+            if not component_to_delete:
+                return jsonify({'status': 'error', 'message': f'Component {data["_id"]} not found'}), 400
+
+            current_order = component_to_delete['order_number']
+
             result = db['Component'].delete_one({'_id': ObjectId(data['_id'])})
 
             if result.deleted_count > 0:
+                # Shift the components below the deleted one by decrementing their order_number
                 db['Component'].update_many({
-                    'order_number': {'$gte': current_order},
-                    'page_id': data['page_id'],},
-                    {'$inc': {'order_number': -1}})
+                    'order_number': {'$gt': current_order},
+                    'page_id': component_to_delete['page_id']
+                }, {'$inc': {'order_number': -1}})
                 return jsonify({'status': 'success', 'message': f'Component {data["_id"]} deleted successfully'}), 200
             else:
                 return jsonify({'status': 'error', 'message': f'Component {data["_id"]} not found'}), 400
